requirements.txt
Django==2.2.3
django-auth-ldap==2.0.0
djangorestframework==3.9.4
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-ldap==3.2.0
pytz==2019.1
sqlparse==0.3.0
